@@ -108,16 +108,17 @@ private:
     Gtk::ToggleButton* bRecentlySaved[2];
     Gtk::ToggleButton* bTrash;
     Gtk::ToggleButton* bNotTrash;
-    Gtk::ToggleButton* categoryButtons[19];
+    Gtk::ToggleButton* bOriginal;
+    Gtk::ToggleButton* categoryButtons[20];
     Gtk::ToggleButton* exifInfo;
-    sigc::connection bCateg[19];
+    sigc::connection bCateg[20];
     Gtk::Image* iFilterClear, *igFilterClear;
     Gtk::Image* iranked[5], *igranked[5], *iUnRanked, *igUnRanked;
     Gtk::Image* iCLabeled[5], *igCLabeled[5], *iUnCLabeled, *igUnCLabeled;
     Gtk::Image* iEdited[2], *igEdited[2];
     Gtk::Image* iRecentlySaved[2], *igRecentlySaved[2];
     Gtk::Image *iTrashEmpty, *iTrashFull;
-    Gtk::Image *iNotTrash;
+    Gtk::Image *iNotTrash, *iOriginal;
     Gtk::Image *iRefreshWhite, *iRefreshRed;
     Gtk::Image *iLeftPanel_1_Show, *iLeftPanel_1_Hide, *iRightPanel_1_Show, *iRightPanel_1_Hide;
     Gtk::Image *iQueryClear;
