@@ -90,6 +90,22 @@ public:
 
 #pragma GCC diagnostic pop
 
+class ConnectionBlocker
+{
+public:
+    ConnectionBlocker (sigc::connection& connection) : connection (connection)
+    {
+        wasBlocked = connection.block();
+    }
+    ~ConnectionBlocker ()
+    {
+        connection.block(wasBlocked);
+    }
+private:
+    sigc::connection& connection;
+    bool wasBlocked;
+};
+
 /**
  * @brief Glue box to control visibility of the MyExpender's content ; also handle the frame around it
  */
@@ -410,7 +426,8 @@ private:
 
 public:
     MyImageMenuItem (Glib::ustring label, Glib::ustring imageFileName);
-    const RTImage *getImage();
+    const RTImage *getImage () const;
+    const Gtk::Label* getLabel () const;
 };
 
 /**
@@ -562,5 +579,13 @@ public:
         return surface ? surface->get_height() : 0;    // sending back the allocated height
     }
 };
+
+inline void setActiveTextOrIndex (Gtk::ComboBoxText& comboBox, const Glib::ustring& text, int index)
+{
+    comboBox.set_active_text (text);
+
+    if (comboBox.get_active_row_number () < 0)
+        comboBox.set_active (index);
+}
 
 #endif
