@@ -224,8 +224,8 @@ void History::bookmarkSelectionChanged ()
 void History::procParamsChanged (ProcParams* params, ProcEvent ev, Glib::ustring descr, ParamsEdited* paramsEdited)
 {
 
-    // to prevent recursion, we filter out the events triggered by the history
-    if (ev == EvHistoryBrowsed) {
+    // to prevent recursion, we filter out the events triggered by the history and events that should not be registered
+    if (ev == EvHistoryBrowsed || ev == EvMonitorTransform) {
         return;
     }
 
