@@ -21,7 +21,7 @@
 #ifndef _POPUPTOGGLEBUTTON_
 #define _POPUPTOGGLEBUTTON_
 
-#include "popupbutton.h"
+#include <gtkmm/togglebutton.h>
 #include "popupcommon.h"
 
 class PopUpToggleButton : public Gtk::ToggleButton, public PopUpCommon
