@@ -22,8 +22,12 @@ protected:
     CurveEditorGroup* curveEditorGD;
     CurveEditorGroup* curveEditorGDH;
     CurveEditorGroup* curveEditorGH;
+    CurveEditorGroup* curveEditormap;
     Adjuster* str;
     Adjuster* scal;
+    Adjuster* grad;
+    Adjuster* grads;
+    Adjuster* iter;
     Adjuster* neigh;
     Adjuster* gain;
     Adjuster* offs;
@@ -34,10 +38,21 @@ protected:
     Adjuster* grbl;
     Adjuster* gam;
     Adjuster* slope;
+    Adjuster* highlights;
+    Adjuster* h_tonalwidth;
+    Adjuster* shadows;
+    Adjuster* s_tonalwidth;
+    Adjuster* radius;
+
     MyExpander* expsettings;
 
     Gtk::Label* labmdh;
     Gtk::HBox* dhbox;
+    Gtk::HBox* mapbox;
+    Gtk::Label* labmap;
+    Gtk::HBox* viewbox;
+    Gtk::Label* labview;
+
     Gtk::Label* labgam;
     Gtk::HBox* gambox;
     Gtk::Button* neutral;
@@ -46,6 +61,8 @@ protected:
     MyComboBoxText*   retinexMethod;
     MyComboBoxText*   retinexcolorspace;
     MyComboBoxText*   gammaretinex;
+    MyComboBoxText*   mapMethod;
+    MyComboBoxText*   viewMethod;
     Gtk::CheckButton* medianmap;
     double nextmin;
     double nextmax;
@@ -62,8 +79,9 @@ protected:
 
     DiagonalCurveEditor* cdshape;
     DiagonalCurveEditor* cdshapeH;
+    DiagonalCurveEditor* mapshape;
     CurveEditorGroup* transmissionCurveEditorG;
-    sigc::connection retinexMethodConn, neutralconn;
+    sigc::connection retinexMethodConn, neutralconn, mapMethodConn, viewMethodConn;
     sigc::connection retinexColorSpaceConn;
     sigc::connection gammaretinexConn;
     FlatCurveEditor* transmissionShape;
@@ -92,6 +110,8 @@ public:
     void enabledChanged        ();
     void curveChanged          (CurveEditor* ce);
     void retinexMethodChanged();
+    void mapMethodChanged();
+    void viewMethodChanged();
     void retinexColorSpaceChanged();
     void gammaretinexChanged();
     void ColorSpaceUpdateUI();
