@@ -26,7 +26,7 @@ ThumbBrowserEntryBase::ThumbBrowserEntryBase (const Glib::ustring& fname)
       prex(0), prey(0), upperMargin(6), borderWidth(1), textGap(6), sideMargin(8), lowerMargin(8),
       preview(NULL), dispname(Glib::path_get_basename (fname)), buttonSet(NULL), width(0), height(0),
       exp_width(0), exp_height(0), startx(0), starty(0), ofsX(0), ofsY(0), redrawRequests(0),
-      parent(NULL), bbSelected(false), bbFramed(false), bbPreview(NULL), cursor_type(CSUndefined),
+      parent(NULL), original(NULL), bbSelected(false), bbFramed(false), bbPreview(NULL), cursor_type(CSUndefined),
       thumbnail(NULL), filename(fname), shortname(dispname), exifline(""), datetimeline(""),
       selected(false), drawable(false), filtered(false), framed(false), processing(false), italicstyle(false),
       edited(false), recentlysaved(false), updatepriority(false), withFilename(WFNAME_NONE) {}
@@ -565,6 +565,17 @@ bool ThumbBrowserEntryBase::insideWindow (int x, int y, int w, int h)
 {
 
     return !(ofsX + startx > x + w || ofsX + startx + exp_width < x || ofsY + starty > y + h || ofsY + starty + exp_height < y);
+}
+
+std::vector<Glib::RefPtr<Gdk::Pixbuf> > ThumbBrowserEntryBase::getIconsOnImageArea()
+{
+    return std::vector<Glib::RefPtr<Gdk::Pixbuf> >();
+}
+
+void ThumbBrowserEntryBase::getIconSize(int& w, int& h)
+{
+    w = 0;
+    h = 0;
 }
 
 bool ThumbBrowserEntryBase::motionNotify  (int x, int y)
